Rewrite some of setup.sh to be more robust.
Add vars that link to book the path to a config file, and the path OF
the config file itself. Removes having to specify these paths multiple
times.

`remove_files` now checks first to see if the file exists before running
`rm`. `workstation_files` now checks to see if the path exists, and runs
`mkdir` if the path does not exists (assumes `$HOME/.config` exists).