Autostart Dunst and Xautolock
Autostart Dunst as the notification daemon. Currently, Dunst has no
special configuration.

Autostart Xautolock. Xautolock is told to exec `slock` after five
minutes of inactivity, notify me ten seconds before it locks via
`notify-send`, disallow disabling/exiting outside of entering the
password, detect if the machine is sleeping as to not autolock after
wake (such as opening a laptop lid), and to hover in the top-right
corner to disable the locking temporarially.