Fix missing "$" in "$HOME"
`HOME` in the brace expansion needs to include the `$` in order to
retreive the corret value. Without it, it was creating the directory
`HOME` within the same directory as the caller of the script.

This also exposed that the files were not being linked correctly. Need
to investigate for a better solution.