Add Micro-Vim configuration org file
Micro-Vim is a small configuration file for Vim based on
microvimrc[0]. It has a small amount of configuration that improves
QoL without adding any bloat to Vim. Since the main usecase for Vim
for me is in resource restricted environments, keeping it lean is
important.

Changes to existing files include:

 - Ignoring the generated Vim configuration file
 - Updating setup.sh to handle the new file/directory
 - Including a `vim` alias inside fish
 - Removing the NeoVim configuration file

[0]: https://github.com/romainl/microvimrc