Add initial Alacritty Org configuration file
Updated the Git ignore and setup script to manange the new
configuration.

This Org file has two file variables set on the first line. This is to
work-around the Org source blocks automatically indenting the YAML
source to match the surrounding Org text. I usually like this, but the
indentation used are tabs. This breakes the YAML spec and results in
Alacritty not being able to parse the config.