Add some specifications and remove the existing prompt
The specifications added are:

- The username & host name
- The directory path
- The Git branch
- The Git status & state
- The status indicator

The removal of the existing prompt is because I made a mistake when
adding the text alongside some as-of-yet not commited code, resulting
in the added text falling in the center of the existing prompt.
Instead of dealing with it properly, I am just deleting it entirely.

Fight me.